server/requirements.txt
fastapi
python-multipart
python-ldap
motor
uvicorn[standard]
python-jose[cryptography]
pymongo
starlette
pydantic
pytz
schedule~=1.1.0
exchangelib~=4.7.3
pandas~=1.4.2
gunicorn
casbin
rtoml